Add http server and rudimentary home page

+++ params.go
@@ -1,8 +1,8 @@
 package main
 
 type ParamsStr struct {
-	PlaintextPort uint
-	TLSPort       uint
+	PlaintextPort uint64
+	TLSPort       uint64
 	DebugModeTLS  bool
 	FullCert      string
 	PrivCert      string
